requirements.txt
streamlit==1.51.0
google-generativeai==0.8.5
python-dotenv==1.2.1
